Extracted `PathUtils`
